// source/client/src/liveness/nose/NoseChallengeParams.ts
/*
 * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: Apache-2.0
 */

export interface NoseChallengeParams {
  readonly imageHeight: number;
  readonly imageWidth: number;
  readonly areaHeight: number;
  readonly areaLeft: number;
  readonly areaTop: number;
  readonly areaWidth: number;
  readonly noseHeight: number;
  readonly noseLeft: number;
  readonly noseTop: number;
  readonly noseWidth: number;
}